Changing names of the variables and correcting backend fucntions

+++ backend/app/controlers/formation/controlerProfessionalActivities.js
@@ -6,8 +6,8 @@ class ControlerProfessionalActivity {
 
     routes() {
         return {
-            base: '/app/professionalActivities',
-            baseID: `/app/professionalActivities/:id`
+            base: '/app/formation/professional_activities',
+            baseID: `/app/formation/professional_activities/:id`
         }
     }
 
@@ -19,10 +19,10 @@ class ControlerProfessionalActivity {
             } else {
                 const content = { ...req.body }
                 if (req.file !== undefined) {
-                    content['pathArquivo'] = req.file.path
+                    content['filePath'] = req.file.path
                 }
-                daoProfessionalActivities.addProfessionalActivity(req.user, content)
-                    .then(resultado => resp.json(resultado))
+                daoProfessionalActivities.add(req.user, content)
+                    .then(result => resp.json(result))
                     .catch(erro => resp.json(erro))
             }
         }
@@ -30,17 +30,28 @@ class ControlerProfessionalActivity {
 
     getProfessionalActivities() {
         return function (req, resp) {
-            daoProfessionalActivities.getProfessionalActivities(req.user)
-                .then(resultado => resp.json(resultado))
+            daoProfessionalActivities.getAll(req.user)
+                .then(result => resp.json(result))
                 .catch(erro => resp.json(erro))
         }
     }
 
     removeProfessionalActivity() {
         return function (req, resp) {
-            daoProfessionalActivities.removeProfessionalActivity(req.user.email, req.params.id)
-                .then(resultado => resp.json(resultado))
-                .catch(erro => resp.json(erro))
+            daoProfessionalActivities.get(req.user, req.params.id)
+                .then(result => {
+                    if (result !== undefined) {
+                        const data = result['formation']['professionalActivities']
+                        if (!((data['filePath'] === undefined) || (data['filePath'] === ''))) {
+                            fs.unlink(`./${data['filePath']}`, err => { console.log('===========\n==========\n===========', err) })
+                        }
+                    }
+                    daoProfessionalActivities.remove(req.user.email, req.params.id)
+                        .then(result => resp.json(result))
+                        .catch(erro => resp.json(erro))
+                })
+                .catch(erro => resp.json({ erro }))
+
         }
     }
 
@@ -51,22 +62,28 @@ class ControlerProfessionalActivity {
                 resp.json({ erros: errosVal })
             } else {
                 const content = { ...req.body }
-                if (req.file !== undefined) {
-                    content['pathArquivo'] = req.file.path
+                if (req.file !== undefined || req.file === '') {
+                    content['filePath'] = req.file.path
                 } else {
-                    delete content['comprovante']
+                    delete content['proof']
                 }
-                daoProfessionalActivities.getProfessionalActivity(req.user, req.body._id)
-                    .then(resultado => {
+                daoProfessionalActivities.get(req.user, req.body._id)
+                    .then(result => {
                         let oldPath = undefined
-                        if (resultado[0] !== undefined) {
-                            if ((resultado[0]['apresentacoes']['pathArquivo'] !== undefined)) {
-                                oldPath = resultado[0]['apresentacoes']['pathArquivo']
+                        if (result !== undefined) {
+                            console.log('Imprimindo o results!= =====================')
+                            const data = result['formation']['professionalActivities']
+                            if ((data['filePath'] !== undefined)) {
+                                oldPath = data['filePath']
+                                if ((content['filePath'] === undefined) || (content['filePath'] === '')) {
+                                    content['filePath'] = oldPath
+                                } else {
+                                    fs.unlink(`./${oldPath}`, err => { console.log('===========\n==========\n===========', err) })
+                                }
                             }
                         }
-                        fs.unlink(`./${oldPath}`, err => { console.log('===========\n==========\n===========', err) })
-                        daoProfessionalActivities.updateProfessionalActivity(content)
-                            .then(resultado => resp.json(resultado))
+                        daoProfessionalActivities.update(content)
+                            .then(result => resp.json(result))
                             .catch(erro => resp.json(erro))
                     })
                     .catch(erro => resp.json({ erro }))
